Move to utils classes
